requirements.txt
# Core data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web scraping and HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
lxml>=4.9.0

# NSE data
nsetools>=1.0.11
nsepy>=0.8

# Zerodha KiteConnect API
kiteconnect>=4.0.0

# Yahoo Finance and other financial data
yfinance>=0.2.18
alpha-vantage>=2.3.1

# Database and storage
sqlite3

# Configuration and environment
python-dotenv>=1.0.0
configparser>=5.3.0

# Utilities
schedule>=1.2.0
python-dateutil>=2.8.0
pytz>=2023.3

# Logging
loguru>=0.7.0

# Data validation
pydantic>=2.4.0

# Optional: Jupyter notebooks for analysis
jupyter>=1.0.0
ipykernel>=6.25.0
